requirements.txt
python-opencv
Pillow
numpy